vista de errores

+++ assets/functions/delete_pokemon.php
@@ -1,17 +1,32 @@
 <?php
     require("db_functions.php");
 
-    try{
-        if(isset($_POST['identificador'])){
-            $identificador = $_POST['identificador'];
-            $img=ejecutar_query("SELECT img FROM pokemon WHERE identificador = $identificador",true)[0]->img;
+    if(isset($_POST['identificador'])){
+        $identificador = $_POST['identificador'];
+        $_img=ejecutar_query("SELECT img FROM pokemon WHERE identificador = $identificador",true);
+        if($_img){
+            $img=$_img[0]->img;
             if(ejecutar_query("DELETE FROM pokemon WHERE identificador = $identificador")){
                 unlink("../../uploads/$img");
                 header("Location: ../../vistas/busqueda.php");
                 exit();
-            }else throw new error ("Error al eliminar pokémon de la db");
-        }
-    }catch(Exception $e){
-        throw new error("Error al eliminar pokémon");
+            }else{
+                session_start();
+                $_SESSION['error']=("Error al eliminar pokémon");
+                header("Location: ../../vistas/error.php");
+                exit();
+            }
+        }else{
+            session_start();
+            $_SESSION['error']=("Error al eliminar pokémon");
+            header("Location: ../../vistas/error.php");
+            exit();
+       }
+    }else{
+        session_start();
+        $_SESSION['error']=("Error no existen datos del pokémon");
+        header("Location: ../../vistas/error.php");
+        exit();
     }
+  
 ?>
